requirements.txt
Django>=1.6.5
django_hstore
djangorestframework==2.3.14